removed chai assertion error, it was not recieving a status in the request in the form the test was expecting. changed server.js with urlparser and user create index (line 78)
