+++ test/test-users.js
@@ -484,4 +484,4 @@ describe('/api/user', function () {
       // });
     });
   });
-});
+
